test(datetime): migrate prefer-wheel
