+++ core/src/components/datetime/test/prefer-wheel/datetime.e2e.ts-snapshots/datetime-wheel-date-time-diff-ios-ltr-Mobile-Firefox-linux.png
