+++ core/src/components/datetime/test/prefer-wheel/datetime.e2e.ts-snapshots/datetime-wheel-time-date-diff-md-ltr-Mobile-Safari-linux.png
